IContextMenuPresenter定义在presenter.d.ts
